decoder: Fix end-of-contents octet handling
X.690 says, "The end-of-contents octets shall consist of two zero
octets." While it "can be considered as the encoding of" a thing with
type 0, other ways of encoding that same thing aren't valid
end-of-contents octets.

Furthermore, end-of-contents octets aren't valid in all contexts, only
in constructed encodings with indefinite length. Previously,
applications could attempt to parse to an asn1Spec and receive the
`endOfOctets` sentinel value: this change makes end-of-contents octets
an error at top level, as well as in recursive decoding of
definite-length encodings.

This bug was originally spotted by Alex Gaynor using the AFL fuzzer: see
https://github.com/pyca/cryptography/issues/1838